+++ src/main/java/uk/gov/hmcts/reform/jobscheduler/services/jobs/SchedulerExceptionHandlingHelper.java
@@ -3,14 +3,11 @@ package uk.gov.hmcts.reform.jobscheduler.services.jobs;
 import org.quartz.SchedulerException;
 import uk.gov.hmcts.reform.jobscheduler.services.jobs.exceptions.JobException;
 
-public class GetterFromScheduler {
+public class SchedulerExceptionHandlingHelper {
 
     @FunctionalInterface
     public interface ThrowingSupplier<T> {
         T get() throws SchedulerException;
-    }
-
-    private GetterFromScheduler() {
     }
 
     public static <T> T call(ThrowingSupplier<T> throwingSupplier) {
@@ -19,5 +16,8 @@ public class GetterFromScheduler {
         } catch (SchedulerException exc) {
             throw new JobException(exc.getMessage(), exc);
         }
+    }
+
+    private SchedulerExceptionHandlingHelper() {
     }
 }
